Moving to a command based version to extract data from the TT
This is not quite ready as there are some aspects of the UCI::pv output
that I haven't been able to reproduce nor have I have been able to
discount as not useful in this context.  On the other hand, this works
well.

Also, probably deserves to be implemented somewhere else. But maybe not.